requirements.txt
solana-py==0.30.2
solders==0.18.1
anchorpy==0.18.0
python-dotenv==1.0.0
pyTelegramBotAPI==4.14.0
